# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Gin + Vue3 的后台管理系统 API。\u003cbr/\u003e统一响应格式：\u003ccode\u003e{\"code\":0,\"msg\":\"ok\",\"data\":...}\u003c/code\u003e，code=0 成功，code=7 业务失败。\u003cbr/\u003e所有 \u003ccode\u003e/api/v1/system/*\u003c/code\u003e 接口需要 \u003ccode\u003eAuthorization: Bearer \u0026lt;token\u0026gt;\u003c/code\u003e。",
        "title": "go-admin API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8989",
    "basePath": "/",
    "paths": {
...
        "/api/v1/system/menu": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "更新菜单",
                "parameters": [
                    {
                        "description": "菜单信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.MenuUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "新增菜单",
                "parameters": [
                    {
                        "description": "菜单信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.MenuCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysMenu"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/system/menu/tree": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "获取全量菜单树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.MenuTreeResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/system/menu/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "删除菜单（递归删除子节点）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                        }
                    }
                }
            }
        },
        "/api/v1/system/role": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "更新角色",
                "parameters": [
                    {
                        "description": "角色信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.RoleUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "新增角色",
                "parameters": [
                    {
                        "description": "角色信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.RoleCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysRole"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/system/role/auth": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "设置角色的菜单/API 权限",
                "parameters": [
                    {
                        "description": "授权参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.RoleAuthReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                        }
                    }
                }
            }
        },
        "/api/v1/system/role/auth/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "查询角色已分配的菜单/API",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.RoleAuthDetailResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/system/role/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "角色列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.RoleListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/system/role/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "删除角色",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                        }
                    }
                }
            }
        },
        "/api/v1/system/role/{id}/default-router": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "设置角色默认首页路由",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "默认路由",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_dto_system.RoleSetDefaultRouterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_utils_response.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_lihongsheng_go-admin_server_dto_system.MenuCreateReq": {
            "type": "object",
            "properties": {
                "api_rules": {
                    "type": "string"
                },
                "component": {
                    "type": "string"
                },
                "hidden": {
                    "type": "boolean"
                },
                "icon": {
                    "type": "string"
                },
                "keep_alive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "permission": {
                    "type": "string"
                },
                "redirect": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "description": "catalog / menu / button，空字符串表示 menu",
                    "type": "string"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_dto_system.MenuTreeResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysMenu"
                    }
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_dto_system.MenuUpdateReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "api_rules": {
                    "type": "string"
                },
                "component": {
                    "type": "string"
                },
                "hidden": {
                    "type": "boolean"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "keep_alive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "permission": {
                    "type": "string"
                },
                "redirect": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_dto_system.RoleAuthDetailResp": {
            "type": "object",
            "properties": {
                "default_router": {
                    "type": "string"
                },
                "menu_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_dto_system.RoleAuthReq": {
            "type": "object",
            "required": [
                "role_id"
            ],
            "properties": {
                "api_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "menu_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "role_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_dto_system.RoleCreateReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "default_router": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_dto_system.RoleListResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysRole"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_dto_system.RoleSetDefaultRouterReq": {
            "type": "object",
            "properties": {
                "default_router": {
                    "type": "string"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_dto_system.RoleUpdateReq": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "default_router": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_lihongsheng_go-admin_server_model_system.SysMenu": {
            "type": "object",
            "properties": {
                "api_rules": {
                    "description": "JSON 数组：[{\"path\":\"/api/...\",\"method\":\"GET\"},...]",
                    "type": "string"
                },
                "children": {
                    "description": "虚拟字段，仅用于 JSON 序列化时组装树",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysMenu"
                    }
                },
                "component": {
                    "description": "前端组件路径；catalog 用 \"Layout\"",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "hidden": {
                    "description": "是否隐藏",
                    "type": "boolean"
                },
                "icon": {
                    "description": "图标名（element-ui icon）",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "keep_alive": {
                    "description": "是否缓存页面",
                    "type": "boolean"
                },
                "name": {
                    "description": "路由 name；button 为按钮名",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父节点 ID，0 表示根",
                    "type": "integer"
                },
                "path": {
                    "description": "路由路径；catalog 如 /system，menu 如 user",
                    "type": "string"
                },
                "permission": {
                    "description": "权限码；button 必有值，如 user:add",
                    "type": "string"
                },
                "redirect": {
                    "description": "重定向路径",
                    "type": "string"
                },
                "sort": {
                    "description": "排序序号",
                    "type": "integer"
                },
                "title": {
                    "description": "显示标题",
                    "type": "string"
                },
                "type": {
                    "description": "catalog / menu / button",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_lihongsheng_go-admin_server_model_system.SysRole": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "default_router": {
                    "description": "登录后默认首页路由",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_lihongsheng_go-admin_server_model_system.SysMenu"
                    }
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_lihongsheng_go-admin_server_utils_response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}